LayoutTests/inspector/css/pseudo-element-matches-for-pseudo-element-node-expected.txt

Testing that we can get computed and matched rules for pseudo element nodes.

TEST ELEMENT: Has Before Pseudo Element
PASS: Got ::before pseudo element for #x
Refreshing styles for ::before pseudo element

PASS: No Inline Styles
PASS: No Attribute Styles
PASS: No Inherited Styles
PASS: No Pseudo Element Styles
PASS: Has Computed Styles
PROPERTY: content: before;
PROPERTY: color: rgb(0, 0, 255);
MATCHED RULE:
  SELECTOR: input:before
  SELECTOR: #x:before [matched]
  SELECTOR: div
